+++ src/libs/Detach/CHANGELOG.md
@@ -10,8 +10,11 @@ This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - Added `Pyramid` 3D primitive along with the following methods:
     - `Geometry3D.SpherePyramid`
     - `Geometry3D.SphereCastPyramid`
+    - `Geometry3D.ClosestPointInPyramid`
   - Added `OrientedPyramid` 3D primitive along with the following methods:
     - `Geoemtry3D.SphereOrientedPyramid`
+    - `Geometry3D.SphereCastOrientedPyramid`
+    - `Geometry3D.ClosestPointInOrientedPyramid`
   - Added `CircleCast` 2D primitive along with the following methods:
     - `Geometry2D.CircleCastPoint`
     - `Geometry2D.CircleCastLine`